requirements.txt
asgiref==3.3.0
Django==3.1.3
djangorestframework==3.12.1
joblib==0.17.0
lightgbm==3.0.0
Markdown==3.3.3
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
